Split account list out into separate component
